+++ packages/create-moralis-dapp/express/package.json
@@ -12,6 +12,8 @@
   ],
   "bin": "./lib/index.js",
   "scripts": {
+    "test": "jest --runInBand --detectOpenHandles --forceExit --ci",
+    "test:coverage": "yarn run test --coverage --coverageReporters json-summary",
     "lint": "eslint . --ext .js,.ts,.tsx,jsx",
     "format": "prettier --write 'src/**/*.{js,ts}'",
     "clean": "rm -rf lib && rm -rf tsconfig.tsbuildinfo && rm -rf ./node_modules/.cache/nx",
@@ -24,6 +26,7 @@
     "@create-moralis-dapp/toolkit": "^2.13.0"
   },
   "devDependencies": {
-    "copyfiles": "^2.4.1"
+    "copyfiles": "^2.4.1",
+    "jest": "29.3.1"
   }
 }
