+++ packages/create-moralis-dapp/express/src/AppGenerator/index.ts
@@ -1,0 +1,1 @@
+export * from './ExpressAppGenerator'
