+++ packages/create-moralis-dapp/express/src/generate.ts
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
-import { ExpressAppGenerator } from './ExpressAppGenerator';
+import { ExpressAppGenerator } from './AppGenerator';
 
 const expressAppGenerator = new ExpressAppGenerator();
 expressAppGenerator.generate();
