+++ packages/create-moralis-dapp/next/src/AppGenerator/index.ts
@@ -1,0 +1,1 @@
+export * from './NextAppGenerator';
