+++ packages/create-moralis-dapp/next/src/generate.ts
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
-import { NextAppGenerator } from './NextAppGenerator';
+import { NextAppGenerator } from './AppGenerator';
 
 const nextAppGenerator = new NextAppGenerator();
 nextAppGenerator.generate();
