+++ packages/create-moralis-dapp/next/src/index.ts
@@ -1,1 +1,1 @@
-export * from './NextAppGenerator';
+export * from './AppGenerator/NextAppGenerator';
